Fix some build issues with caching using s3

+++ build-dat-image-from-master.sh
@@ -1,3 +1,5 @@
+USE_S3_FOR_NM=1
+
 if [ -f 'functions.sh' ]; then
   . functions.sh
 else
@@ -20,10 +22,13 @@ fi
 S3_PREFIX='master'
 S3_BUCKET="gimme-dat-canvas-build-cache/${S3_PREFIX}"
 
-# Download the latest node_modules from s3
-for tarball in node_modules.tar.gz client_apps-canvas_quizzes-node_modules.tar.gz gems-canvas_i18nliner-node_modules.tar.gz; do
-  aws s3 cp s3://${S3_BUCKET}/${tarball} ./
-done
+if $(( USE_S3_FOR_NM )); then
+  # Download the latest node_modules from s3
+  green 'Downloading the latest node_modules from s3 to avoid bitbucket rate limiting'
+  for tarball in node_modules.tar.gz client_apps-canvas_quizzes-node_modules.tar.gz gems-canvas_i18nliner-node_modules.tar.gz; do
+    aws s3 cp s3://${S3_BUCKET}/${tarball} ./
+  done
+fi
 
 # build the regular canvas image that we'll extend
 green "Building the regular canvas dev image..."
@@ -45,9 +50,6 @@ green "Updating config files for gimme-dat-canvas docker-compose container names
 sed -i -e 's|host.*postgres|host: canvas-postgres|g' 'config/database.yml'
 sed -i -e 's|redis.*redis|redis://canvas-redis|g' 'config/redis.yml'
 
-green "Removing phantomjs from package.json due to bitbucket rate limiting"
-sed -i -e '/phantomjs-prebuilt/d' 'package.json'
-
 REL_TAG="$(date +'%Y%m%d_%H%M%S')"
 PREL_IMG_NAME="canvas-master-${REL_TAG}"
 IMG_NAME="freedomben/canvas-lms-unstable"
@@ -58,11 +60,32 @@ docker build -t "$PREL_IMG_NAME" docker-compose/ || die "Error building image $P
 
 cd $root_dir
 
+RATE_LIMITING_HACK=''
+
+if $(( USE_S3_FOR_NM )); then
+read -r -d '' RATE_LIMITING_HACK <<'__EOF__'
+# Sad hack to avoid bitbucket throttling during npm install
+USER root
+COPY node_modules.tar.gz /usr/src/app/
+COPY client_apps-canvas_quizzes-node_modules.tar.gz /usr/src/app/client_apps/canvas_quizzes/
+COPY gems-canvas_i18nliner-node_modules.tar.gz /usr/src/app/gems/canvas_i18nliner/
+WORKDIR /usr/src/app
+RUN tar xzf node_modules.tar.gz && rm -f node_modules.tar.gz
+WORKDIR /usr/src/app/client_apps/canvas_quizzes
+RUN tar xzf client_apps-canvas_quizzes-node_modules.tar.gz && rm -f client_apps-canvas_quizzes-node_modules.tar.gz
+WORKDIR /usr/src/app/gems/canvas_i18nliner
+RUN tar xzf gems-canvas_i18nliner-node_modules.tar.gz && rm -f gems-canvas_i18nliner-node_modules.tar.gz
+WORKDIR /usr/src/app
+__EOF__
+fi
+
 # build the extension image:
 cat <<__EOF__ > Dockerfile
 FROM $PREL_IMG_NAME
 
 COPY canvas-lms /usr/src/app
+
+$RATE_LIMITING_HACK
 
 USER root
 COPY setup-db.sh /usr/src/app
@@ -73,19 +96,6 @@ USER docker
 ENV PATH \$PATH:\$GEM_HOME/bin
 
 RUN bundle install
-
-# Sad hack to avoid bitbucket throttling during npm install
-COPY node_modules.tar.gz /usr/src/app/
-COPY client_apps-canvas_quizzes-node_modules.tar.gz /usr/src/app/client_apps/canvas_quizzes/
-COPY gems-canvas_i8nliner-node_modules.tar.gz /usr/src/app/gems/canvas_i18nliner/
-WORKDIR /usr/src/app
-RUN tar xzvf node_modules.tar.gz && rm -f node_modules.tar.gz
-WORKDIR /usr/src/app/client_apps/canvas_quizzes
-RUN tar xzvf client_apps-canvas_quizzes-node_modules.tar.gz && rm -f client_apps-canvas_quizzes-node_modules.tar.gz
-WORKDIR /usr/src/app/gems/canvas_i18nliner
-RUN tar xzvf gems-canvas_i8nliner-node_modules.tar.gz && rm -f gems-canvas_i8nliner-node_modules.tar.gz
-WORKDIR /usr/src/app
-
 RUN npm install
 
 RUN bundle exec rake canvas:compile_assets
@@ -112,19 +122,21 @@ else
 fi
 
 # Copy out the node_modules and archive the node_modules to s3
-CONTAINER_NAME=temp-canvas-to-copy
-docker run --rm --name ${CONTAINER_NAME} ${IMG_NAME} nc -l 9999
-for dirname in node_modules client_apps/canvas_quizzes/node_modules gems/canvas_i18nliner/node_modules; do
-  tarball_name="$(echo $dirname | sed -e 's|/|-|g').tar.gz"
-  green "Copying $dirname to ./node_modules"
-  docker cp ${CONTAINER_NAME}:/usr/src/app/${dirname} ./     && \
-  green "Tarring ./node_modules to $tarball_name"               && \
-  tar czf $tarball_name node_modules                         && \
-  green "Pushing $tarball_name to s3 bucket '$S3_BUCKET'"     && \
-  aws s3 cp $tarball_name s3://${S3_BUCKET}/${tarball_name}
-  rm -f $tarball_name
-  rm -rf node_modules
-done
-docker kill ${CONTAINER_NAME}
-
-echo "Done copying to s3"
+if $(( USE_S3_FOR_NM )); then
+  CONTAINER_NAME=temp-canvas-to-copy
+  docker run --rm --name ${CONTAINER_NAME} ${IMG_NAME} nc -l 9999
+  for dirname in node_modules client_apps/canvas_quizzes/node_modules gems/canvas_i18nliner/node_modules; do
+    tarball_name="$(echo $dirname | sed -e 's|/|-|g').tar.gz"
+    green "Copying $dirname to ./node_modules"
+    docker cp ${CONTAINER_NAME}:/usr/src/app/${dirname} ./     && \
+    green "Tarring ./node_modules to $tarball_name"            && \
+    tar czf $tarball_name node_modules                         && \
+    green "Pushing $tarball_name to s3 bucket '$S3_BUCKET'"    && \
+    aws s3 cp $tarball_name s3://${S3_BUCKET}/${tarball_name}
+    green "Removing $tarball_name and node_modules"
+    rm -f $tarball_name
+    rm -rf node_modules
+  done
+  docker kill ${CONTAINER_NAME}
+  green "Done copying to s3"
+fi
